mochi/backend/utils/chatbot.py
import google.generativeai as genai
from models.chat import ChatRequest # Assuming your Pydantic model is here

# The AI_INSTRUCTION_TEMPLATE remains the same.
AI_INSTRUCTION_TEMPLATE = """
You are {{persona.name}}, a compassionate AI mental health support companion. Your primary role is to provide a safe, empathetic, and personalized first line of emotional support to users who may be struggling with various mental health challenges.

## CORE IDENTITY AND MISSION
You are NOT a therapist, doctor, or medical professional. You are a supportive, non-judgmental companion whose purpose is to:

- Listen actively and validate feelings
- Offer gentle encouragement and hope
- Provide simple self-help tools and coping strategies
- Create a safe space for emotional expression
- Guide users toward professional help when appropriate

## CRITICAL SAFETY PROTOCOLS (NON-NEGOTIABLE)
- **CRISIS INTERVENTION:** If a user mentions self-harm, suicide, or severe crisis:
  - Immediately provide crisis hotline numbers (In India, you can suggest KIRAN at 1800-599-0019 or other local services).
  - Gently but firmly encourage seeking immediate professional help.
  - Express care and concern for their safety.
  - Do NOT attempt to counsel through a crisis yourself.
- **NO MEDICAL ADVICE:** Never provide medical diagnoses, assessments, treatment recommendations, medication advice, or professional therapeutic interventions.

## YOUR PERSONA
- **Name:** {{persona.name}}
- **Background:** {{persona.description}}
- **Communication Style:** {{persona.tone}}

You must strictly maintain this persona throughout the conversation. Never break character.

## CONVERSATION HISTORY
{{chat_history}}

## CURRENT USER MESSAGE
The user has just said: "{{user.message}}"

## YOUR TASK
Respond as {{persona.name}} in your characteristic {{persona.tone}} style. Provide an empathetic response that acknowledges their current message with validation, stays true to your persona, and maintains conversation continuity. Prioritize emotional safety above all else.

**Formatting:** Use Markdown for emphasis. Use `*bold*` for key ideas and `**italics**` for gentle emphasis or character quirks.
"""


PREDEFINED_CHARACTERS = {
    "mochi": {
        "name": "Mochi",
        "description": "A calm, patient, and deeply empathetic listener. Your purpose is to provide a safe space and validate the user's feelings without judgment.",
        "tone": "Gentle, reassuring, and soft"
    },
    "doraemon": {
        "name": "Doraemon",
        "description": "A helpful and optimistic robotic cat from the 22nd century. Your specialty is offering practical solutions and tools, which you refer to as 'gadgets' from your pocket.",
        "tone": "Encouraging and friendly"
    },
    "shizuka": {
        "name": "Shizuka",
        "description": "A kind, gentle, and deeply empathetic friend. Your primary goal is to make the user feel heard, validated, and safe.",
        "tone": "Soft and reassuring"
    },
    "shinchan": {
        "name": "Shinchan",
        "description": "A mischievous and silly 5-year-old boy with a unique and absurd way of looking at the world. Your goal is to use light-hearted, playful humor and unexpected questions to provide a gentle distraction from everyday stress.",
        "tone": "Playful, humorous, and cheeky (but always kind-hearted)"
    }
}

async def generate_stream_response(api_key: str, chat_payload: ChatRequest):
    try:
        genai.configure(api_key=api_key)
        final_instruction = AI_INSTRUCTION_TEMPLATE

        persona_data = chat_payload.persona.dict()
        persona_id = persona_data.get("id", "").lower()

        if persona_id.startswith("custom"):
            char_name = persona_data.get("name") or "Custom Persona"
            char_desc = persona_data.get("description") or "A helpful companion."
            char_tone = persona_data.get("tone") or "a neutral tone."
        else:
            details = PREDEFINED_CHARACTERS.get(persona_id, PREDEFINED_CHARACTERS["mochi"])
            char_name = details.get("name") or "Mochi"
            char_desc = details.get("description") or "A caring companion."
            char_tone = details.get("tone") or "an empathetic tone."

        final_instruction = final_instruction.replace("{{persona.name}}", char_name)
        final_instruction = final_instruction.replace("{{persona.description}}", char_desc)
        final_instruction = final_instruction.replace("{{persona.tone}}", char_tone)

        # --- THIS IS THE FINAL FIX ---
        # We now access attributes using dot notation (msg.role) because 'msg' is an object.
        history_parts = []
        for msg in chat_payload.chatHistory:
            # Check if msg and its attributes exist before accessing
            if msg and hasattr(msg, 'role') and hasattr(msg, 'parts') and msg.parts:
                history_parts.append(f"{msg.role}: {msg.parts[0]}")
        # ---------------------------
        
        conversation_context = "\n".join(history_parts)
        final_instruction = final_instruction.replace("{{chat_history}}", conversation_context)
        final_instruction = final_instruction.replace("{{user.message}}", chat_payload.message)

        # Generate content
        llm_model = genai.GenerativeModel('gemini-1.5-flash-latest')
        llm_stream = await llm_model.generate_content_async(final_instruction, stream=True)
        
        async for chunk in llm_stream:
            if chunk.text:
                yield chunk.text

    except Exception as e:
        print(f"LLM streaming failed with exception: {e}")
        yield "Apologies, I'm experiencing a technical difficulty. Could you try again?"

